Moved densify_poly func to util.py and simplified it using shapely.segmentize. Simplified the length check for the warning in the SpatialAverager
