+++ xesmf/util.py
@@ -2,7 +2,7 @@ import warnings
 
 import numpy as np
 import xarray as xr
-from shapely.geometry import MultiPolygon, Polygon
+from shapely.geometry import MultiPolygon, Polygon, LineString, MultiLineString
 
 LON_CF_ATTRS = {'standard_name': 'longitude', 'units': 'degrees_east'}
 LAT_CF_ATTRS = {'standard_name': 'latitude', 'units': 'degrees_north'}
@@ -211,6 +211,30 @@ def split_polygons_and_holes(polys):
 
     return exteriors, holes, i_ext, i_hol
 
+
+def densify_polys(polys, max_length=1.0):
+    """
+    Takes in a list of polygons and densifies each of the polygons' segments by adding equidistant points
+    between the two endpoints of each segment such that the maximum length of a segment is max_length.
+    max_length is set to 1.0 degree in lon/lat space by default.
+
+    Parameters
+    ----------
+    polys : list, List of polygons to densify
+
+    max_length : float, Maximum desired length of a segment. The default is 1 degree in lat/lon space.
+                        Each side of the polygon is then split into equidistant segments of length max_length
+
+    Returns
+    -------
+    List of densified polygons
+
+    """
+    dense_polys = []  # Densified polys
+    for poly in polys:
+        dense_poly = poly.segmentize(max_segment_length=max_length)
+        dense_polys.append(dense_poly)
+    return dense_polys
 
 # Constants
 PI_180 = np.pi / 180.0
